requirements.txt
aiogram==3.4.1
asyncpg==0.29.0
aiofiles==23.2.1
python-dotenv==1.0.1
pytz==2024.1
aiohttp==3.9.5
psutil==5.9.8
